feat(DTFS2-6023): Facility amendment SOFs
