+++ azure-functions/acbs-function/acbs-amend-facility-loan-record/function.json
@@ -1,0 +1,9 @@
+{
+    "bindings": [
+        {
+            "name": "context",
+            "type": "orchestrationTrigger",
+            "direction": "in"
+        }
+    ]
+}
