+++ azure-functions/acbs-function/acbs-amend-facility-master-record/function.json
@@ -1,0 +1,9 @@
+{
+    "bindings": [
+        {
+            "name": "context",
+            "type": "orchestrationTrigger",
+            "direction": "in"
+        }
+    ]
+}
