+++ azure-functions/acbs-function/acbs-amend-facility/index.js
@@ -20,10 +20,6 @@
 
 const df = require('durable-functions');
 const retryOptions = require('../helpers/retryOptions');
-const mappings = require('../mappings');
-const CONSTANTS = require('../constants');
-
-const { FACILITY } = CONSTANTS;
 
 module.exports = df.orchestrator(function* amendACBSFacility(context) {
   try {
@@ -46,107 +42,39 @@ module.exports = df.orchestrator(function* amendACBSFacility(context) {
         const { acbsFacility: fmr, etag } = yield context.df.callActivityWithRetry('activity-get-facility-master', retryOptions, { facilityId });
 
         if (!!fmr && !!etag && !!facility.facilitySnapshot) {
-          let facilityMasterRecordAmendments;
-          let facilityLoanRecordAmendments;
           const amendments = {
             amendment,
           };
-          const { facilitySnapshot } = facility;
 
           /**
-           * *************************** AMENDMENT MAPPING ***************************
+           * *************************** AMENDMENT SOFs ***************************
            */
 
-          // 2.1. Facility Master Record (FMR)
-          const fmrMapped = mappings.facility.facilityMasterAmend(fmr, amendments, deal);
-
-          // 2.2. FMR update
-          // 2.2.1 - UKEF Exposure
-          if (amendment.amount) {
-            const result = yield context.df.callActivityWithRetry('activity-update-facility-master', retryOptions, {
-              facilityId,
-              acbsFacilityMasterInput: fmrMapped,
-              updateType: 'amendAmount',
-              etag,
-            });
-
-            facilityMasterRecordAmendments = {
-              amount: {
-                ...result,
-              },
-            };
-          }
-
-          // 2.2.2 - Cover end date
-          if (amendment.coverEndDate) {
-            // 2.2.3. DAF : activity-get-facility-master: Retrieve ACBS `Facility Master Record` with new eTag
-            const updatedFmr = yield context.df.callActivityWithRetry('activity-get-facility-master', retryOptions, { facilityId });
-
-            if (updatedFmr.etag) {
-              const result = yield context.df.callActivityWithRetry('activity-update-facility-master', retryOptions, {
-                facilityId,
-                acbsFacilityMasterInput: fmrMapped,
-                updateType: 'amendExpiryDate',
-                etag: updatedFmr.etag,
-              });
-
-              facilityMasterRecordAmendments = {
-                ...facilityMasterRecordAmendments,
-                coverEndDate: {
-                  ...result,
-                },
-              };
-            }
-          }
-
-          // 2.3. Facility Loan Record (FLR)
-          const flrMApped = mappings.facility.facilityLoanAmend(amendments, facility, fmr);
-
-          // 2.3 FLR update
-          // 2.3.1 Extract loan id for facility id
-          const loanId = yield context.df.callActivityWithRetry('activity-get-loan-id', retryOptions, {
+          // 1. SOF: Facility Loan Record (FLR)
+          const facilityLoanRecord = context.df.callSubOrchestrator('acbs-amend-facility-loan-record', {
             facilityId,
+            facility,
+            amendments,
+            fmr,
           });
 
-          if (loanId) {
-            // 2.3.2 - UKEF Exposure
-            // Loan amount - All except `Loan` facility type
-            if (amendment.amount && facilitySnapshot.type !== FACILITY.FACILITY_TYPE.LOAN) {
-              const result = yield context.df.callActivityWithRetry('activity-update-facility-loan-amount', retryOptions, {
-                loanId,
-                facilityId,
-                acbsFacilityLoanInput: flrMApped,
-              });
+          yield context.df.Task.all([facilityLoanRecord]);
 
-              facilityLoanRecordAmendments = {
-                amount: {
-                  ...result,
-                },
-              };
-            }
+          // 2. SOF: Facility Master Record (FMR)
+          const facilityMasterRecord = context.df.callSubOrchestrator('acbs-amend-facility-master-record', {
+            deal,
+            facilityId,
+            fmr,
+            etag,
+            amendments,
+          });
 
-            // 2.3.3 - Cover end date
-            if (amendment.coverEndDate) {
-              const result = yield context.df.callActivityWithRetry('activity-update-facility-loan', retryOptions, {
-                loanId,
-                facilityId,
-                acbsFacilityLoanInput: flrMApped,
-              });
-
-              facilityLoanRecordAmendments = {
-                ...facilityLoanRecordAmendments,
-                coverEndDate: {
-                  ...result,
-                },
-              };
-            }
-          }
+          yield context.df.Task.all([facilityMasterRecord]);
 
           return {
             facilityId,
-            loanId,
-            facilityMasterRecordAmendments,
-            facilityLoanRecordAmendments,
+            facilityLoanRecord: facilityLoanRecord?.result,
+            facilityMasterRecord: facilityMasterRecord?.result,
           };
         }
         throw new Error('ACBS facility amendment error : Unable to retrieve FMR.');
